Split core/test sources into 2.11 and 2.12 specific code
